fix: убрал закоменченные строки, поменял формат счетчика жалоб, поправил обновление статуса

+++ ssai-main/ssai-main-server/src/main/java/ru/softplat/main/server/model/product/Product.java
@@ -82,5 +82,5 @@ public class Product {
     Float rating;
 
     @Column(name = "complaint_count")
-    Long complaintCount;
+    Integer complaintCount;
 }
